Update terms I use to reference myself
Engineer - I did not complete my degree, so I'm not deserving of this title, or at least it is not appropriate.
Seasoned - I'm proficient, but I've been doing this for just three years at this point. Using the term "seasoned" feels dishonest, as I did not clearly understand that word before using it.